Create separate `settingeditor` and `settingeditor-extension` packages.
