+++ jupyterlab/package.json
@@ -51,6 +51,7 @@
     "@jupyterlab/running": "^0.12.0",
     "@jupyterlab/running-extension": "^0.12.0",
     "@jupyterlab/services": "^0.51.0",
+    "@jupyterlab/settingeditor": "^0.1.0",
     "@jupyterlab/settingeditor-extension": "^0.7.0",
     "@jupyterlab/shortcuts-extension": "^0.12.0",
     "@jupyterlab/tabmanager-extension": "^0.12.0",
@@ -178,6 +179,7 @@
       "@jupyterlab/running": "../packages/running",
       "@jupyterlab/running-extension": "../packages/running-extension",
       "@jupyterlab/services": "../packages/services",
+      "@jupyterlab/settingeditor": "../packages/settingeditor",
       "@jupyterlab/settingeditor-extension": "../packages/settingeditor-extension",
       "@jupyterlab/shortcuts-extension": "../packages/shortcuts-extension",
       "@jupyterlab/tabmanager-extension": "../packages/tabmanager-extension",
