+++ packages/metapackage/package.json
@@ -32,6 +32,7 @@
     "watch:files": "node watch-files.js"
   },
   "dependencies": {
+<<<<<<< HEAD
     "@jupyterlab/application": "^0.12.0",
     "@jupyterlab/application-extension": "^0.12.0",
     "@jupyterlab/apputils": "^0.12.4",
@@ -73,6 +74,7 @@
     "@jupyterlab/running": "^0.12.0",
     "@jupyterlab/running-extension": "^0.12.0",
     "@jupyterlab/services": "^0.51.0",
+    "@jupyterlab/settingeditor": "^0.1.0",
     "@jupyterlab/settingeditor-extension": "^0.7.0",
     "@jupyterlab/shortcuts-extension": "^0.12.0",
     "@jupyterlab/tabmanager-extension": "^0.12.0",
