+++ packages/settingeditor/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@jupyterlab/settingeditor-extension",
-  "version": "0.7.0",
-  "description": "JupyterLab - Setting Editor Extension",
+  "name": "@jupyterlab/settingeditor",
+  "version": "0.1.0",
+  "description": "The JupyterLab default setting editor interface",
   "homepage": "https://github.com/jupyterlab/jupyterlab",
   "bugs": {
     "url": "https://github.com/jupyterlab/jupyterlab/issues"
@@ -26,15 +26,12 @@
   "scripts": {
     "build": "tsc",
     "clean": "rimraf lib",
-    "prepublishOnly": "npm run build",
     "watch": "tsc -w"
   },
   "dependencies": {
-    "@jupyterlab/application": "^0.12.0",
-    "@jupyterlab/apputils": "^0.12.4",
-    "@jupyterlab/codeeditor": "^0.12.0",
-    "@jupyterlab/coreutils": "^0.12.0",
-    "@jupyterlab/settingeditor": "^0.1.0",
+    "@jupyterlab/apputils": "^0.11.1",
+    "@jupyterlab/codeeditor": "^0.11.1",
+    "@jupyterlab/coreutils": "^0.11.1",
     "@phosphor/coreutils": "^1.3.0",
     "@phosphor/messaging": "^1.2.2",
     "@phosphor/signaling": "^1.2.2",
@@ -44,8 +41,5 @@
   "devDependencies": {
     "rimraf": "~2.6.2",
     "typescript": "~2.4.2"
-  },
-  "jupyterlab": {
-    "extension": true
   }
 }
