+++ packages/settingeditor/src/plugineditor.ts
