+++ packages/settingeditor/src/pluginlist.ts
