+++ packages/settingeditor/src/splitpanel.ts
